[COD-34][WIP] Add service for session refresh
